support handling root classes
These changes started out as something simple: get an NSArray with the
root classes.  Unfortunately, collections, including NSArray, crash if
they can't call -retain on their contents, so you can't make an NSArray
of the root classes, because some of them do not implement the NSObject
protocol. rdar://8487380  This is a not an issue in a garbage-collected
environment, since objc_msgsend() eats calls to -retain, but we clearly
need to support non-garbage-collected runtimes.

So, we need a wrapper for the Class type.  Then, the issue was what to
do about metaclasses.  Even the C API seems to be of two minds about
the degree to which metaclasses are an implementation detail, e.g.
there are separate class_getClass*() and classGetInstance*() functions,
but only one class_copy*List(), which only returns things defined on
the normal class, not the metaclass.  I talked with @rentzsch and
@jemmons, and we agreed that Objective-C coders, even if they know
about metaclasses, don't think in terms of them, e.g. there is *one*
conceptual NSObject class, and it has both instance and class methods.